refactor: 존재하는지 확인하는 로직을 exsit로 변경
