+++ src/main/java/roomescape/reservation/repository/ReservationRepository.java
@@ -118,21 +118,18 @@ public class ReservationRepository {
     }
 
     public boolean existSameDateTime(ReservationDate reservationDate, Long timeId) {
-        String sql = "SELECT COUNT(*) FROM reservation WHERE date = ? AND time_id = ?";
-        int count = jdbcTemplate.queryForObject(sql, Integer.class, reservationDate.getDate(), timeId);
-        return count > 0;
+        String sql = "SELECT EXISTS(SELECT 1 FROM reservation WHERE date = ? AND time_id = ?)";
+        return jdbcTemplate.queryForObject(sql, boolean.class, reservationDate.getDate(), timeId);
     }
 
     public boolean existReservationByTimeId(Long timeId) {
-        String sql = "SELECT COUNT(*) FROM reservation WHERE time_id = ?";
-        int count = jdbcTemplate.queryForObject(sql, Integer.class, timeId);
-        return count > 0;
+        String sql = "SELECT EXISTS(SELECT 1 FROM reservation WHERE time_id = ?)";
+        return jdbcTemplate.queryForObject(sql, boolean.class, timeId);
     }
 
     public boolean existReservationByThemeId(Long themeId) {
-        String sql = "SELECT COUNT(*) FROM reservation WHERE theme_id = ?";
-        int count = jdbcTemplate.queryForObject(sql, Integer.class, themeId);
-        return count > 0;
+        String sql = "SELECT EXISTS(SELECT 1 FROM reservation WHERE theme_id = ?)";
+        return jdbcTemplate.queryForObject(sql, boolean.class, themeId);
     }
 
 }
